Move `augmented_chain.py` and `prev_transaction_block.py` to `chia/consensus` (#19587)
