+++ chia/_tests/blockchain/test_augmented_chain.py
@@ -10,8 +10,8 @@ from chia_rs.sized_ints import uint32
 
 from chia._tests.blockchain.blockchain_test_utils import _validate_and_add_block
 from chia._tests.util.blockchain import create_blockchain
+from chia.consensus.augmented_chain import AugmentedBlockchain
 from chia.simulator.block_tools import BlockTools
-from chia.util.augmented_chain import AugmentedBlockchain
 from chia.util.errors import Err
