+++ chia/simulator/add_blocks_in_batches.py
@@ -5,12 +5,12 @@ from typing import Optional
 from chia_rs import FullBlock
 from chia_rs.sized_ints import uint32
 
+from chia.consensus.augmented_chain import AugmentedBlockchain
 from chia.consensus.block_body_validation import ForkInfo
 from chia.consensus.difficulty_adjustment import get_next_sub_slot_iters_and_difficulty
 from chia.full_node.full_node import FullNode, PeakPostProcessingResult
 from chia.types.peer_info import PeerInfo
 from chia.types.validation_state import ValidationState
-from chia.util.augmented_chain import AugmentedBlockchain
 from chia.util.batches import to_batches
